Rename tiktak to multistart.
