+++ src/optimagic/optimization/multistart.py
